add module function
