+++ scripts/main.js
@@ -1,3 +1,4 @@
+import {inputImage, inputEdit, editContent, popupContent, validateForm, validateLink} from './function.js';
 import CardList from './cardlist.js';
 import Card from './card.js';
 import Popup from './popup.js';
@@ -13,80 +14,20 @@ const editForm = document.forms.editForm;
 editForm.elements.name.value = document.querySelector('.user-info__name').textContent;
 editForm.elements.link.value = document.querySelector('.user-info__job').textContent;
 const serverUrl = NODE_ENV === 'development' ? 'http://praktikum.tk/cohort2' : 'https://praktikum.tk/cohort2'
-
 const popupNew = new Popup(popup);
-const popupEditNew = new Popup(edit);
+export const popupEditNew = new Popup(edit);
 const popupImage = new Popup(imgPopup);
-
-function inputImage (event) {
-  if (event.target.classList.contains('.place-card__image')) {
-    const img = event.target.closest('.place-card__image');
-    const imgmore = document.querySelector('.imgpopup__image');
-    const imgDelate = img.style.backgroundImage.split('"');
-    imgmore.src = imgDelate[1];
-  }
+const userInfo = {
+  baseUrl: serverUrl,
+  headers: {
+    authorization: 'e6f04cd4-4d82-48d7-8c34-111d7888cce2',
+    'Content-Type': 'application/json'
+  }  
 }
 
+export const api = new Api(userInfo);
+
 placesList.addEventListener('click', inputImage);
-
-function inputEdit (editName, editJob) {
-  const name = document.querySelector('.user-info__name');
-  const job = document.querySelector('.user-info__job');
-
-  name.textContent = editName;
-  job.textContent = editJob;
-};
-
-function editContent (event) {
-  const name = editForm.elements.name.value;
-  const job = editForm.elements.link.value;
-  if (editForm.elements.name.value.length !== 0 && editForm.elements.link.value.length !== 0) {
-    api.addUserInfo(name, job)
-        .then((user) => {
-          document.querySelector('.user-info__name').textContent = user.name;
-          document.querySelector('.user-info__job').textContent = user.about;
-          document.querySelector('.user-info__photo').setAttribute('style', `background-image: url(${user.avatar})`);
-          popupEditNew.close();
-        })
-        .catch((err) => {
-          console.log(err); 
-        });
-    event.preventDefault();
-    popupEditNew.close();
-  }
-};
-
-function popupContent (event) {
-  const imageName = popupForm.elements.name.value;
-  const imageLink = popupForm.elements.link.value;
-  if (popupForm.elements.name.value.length !== 0 && popupForm.elements.link.value.length !== 0) {
-    newCard.addCard(imageLink, imageName);
-    event.preventDefault();
-    popup.classList.toggle('popup_is-opened');
-    popupForm.elements.name.value = '';
-    popupForm.elements.link.value = '';
-  }
-};
-
-function validateForm (event) {
-  const errorSelector = event.target.nextElementSibling;
-  if (event.target.validity.typeMismatch && event.target.type === 'url') {
-    errorSelector.textContent = 'Здесь должна быть ссылка';
-  } else if (event.target.value.length === 0) {
-    errorSelector.textContent = 'Это обязательное поле';
-  } else if (event.target.value.length < 2 || event.target.value.length > 30) {
-    errorSelector.textContent = 'Должно быть от 2 до 30 символов';
-  } else {
-    errorSelector.textContent = '';
-  }
-};
-
-function validateLink (event) {
-  const errorSelector = event.target.nextElementSibling;
-  if (event.target.value.length > 30) {
-    errorSelector.textContent = '';
-  }
-};
 
 root.addEventListener('click', function(event) {
   if(event.target.classList.contains('user-info__button')) {
@@ -104,16 +45,6 @@ popup.addEventListener('submit', popupContent);
 popupForm.addEventListener('input', validateForm);
 popupForm.addEventListener('input', validateLink);
 
-const userInfo = {
-  baseUrl: serverUrl,
-  headers: {
-    authorization: 'e6f04cd4-4d82-48d7-8c34-111d7888cce2',
-    'Content-Type': 'application/json'
-  }  
-}
-
-const api = new Api(userInfo);
-
 api.getUserInfo()
   .then(user => {
     document.querySelector('.user-info__name').textContent = user.name;
@@ -122,7 +53,7 @@ api.getUserInfo()
   })
     .catch((err) => {
     console.log(err); 
-    });
+});
   
 
 api.getInitialCards()
@@ -131,7 +62,7 @@ api.getInitialCards()
   })
     .catch((err) => {
     console.log(err); 
-    });
+});
